requirements.txt
# ============================================
# ОСНОВНЫЕ БИБЛИОТЕКИ ДЛЯ ETL
# ============================================
# Работа с данными
pandas>=2.0.0
numpy>=1.24.0
# HTTP запросы к API
requests>=2.31.0
python-dotenv>=1.0.0
# Базы данных
psycopg2-binary>=2.9.0
sqlalchemy>=1.4,<2.0 
# ============================================
# Airflow (для Python 3.13 и ниже)
# ============================================
# apache-airflow>=2.8.0  # Раскомментировать если Python < 3.14
# ============================================
# ДАШБОРД И ВИЗУАЛИЗАЦИЯ
# ============================================
streamlit>=1.35.0
plotly>=5.24.0
# ============================================
# УТИЛИТЫ
# ============================================
tqdm>=4.66.0
# ============================================
# СОПОСТАВЛЕНИЕ АРТИКУЛОВ (fuzzy matching)
# ============================================
fuzzywuzzy>=0.18.0      # Нечеткое сравнение строк
python-Levenshtein>=0.25.0  # Быстрое вычисление расстояния Левенштейна
# Альтернатива (более современная)
# thefuzz>=0.22.0       # Форк fuzzywuzzy
